requirements.txt
Flask
Flask-MySQLdb
Flask-SQLAlchemy
python-dotenv
PyMySQL
cryptography